Move ASA module update logic to streamAnalyticsManager, update Codelens pattern, remove unnecessary command in package.json
